Refactor WebRTC session into focused modules
